add isl_schedule_node_band_join_child
This function essentially performs the inverse of
isl_schedule_tree_band_split, except that the spaces may
be different since isl_schedule_tree_band_split (in general)
has no option but to introduce anonymous spaces, while
isl_schedule_node_band_join_child can take the product of the spaces.

The permutable and coincident properties are cleared,
because it is impossible to tell whether the members
of the second band are valid at the position of the first band.

The meaning of the isolate option may not be preserved in all cases.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>