add partial/mock isl_system
There are quite a few operations in isl that internally only
operate on a system of linear equality and inequality constraints.
In practice, they operate on an isl_basic_set,
but ignore the space and local variable information.
Introduce an explicit type for use in such circumstances.
For now, the isl_system is implemented in terms of
the isl_basic_map type, but in future, the constraint handling
part of an isl_basic_map could be extracted out to
an independent isl_system type.

Reviewed-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>