export isl_union_set_list
The isl_union_set_list type is needed to be able to export functions
such as isl_schedule_node_insert_sequence.

Only methods that correspond directly to isl functions are exported
in this commit.
It may be useful to add support for range-based for loops
in the C++ bindings in a future commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>