isl_tab_basic_set_constrained_lexmin: allow constraints to be marked optional
An optional constraint is treated in the same way as any other constraint
except that when all the cases have failed at a certain level
of the search corresponding to such a constraint, an extra case
is considered where the constraint is not enforced.
The constraint is also disabled in this subtree of the search space.

The optional constraints currently do not affect the optimality
of the solution.  That is, as soon as all initial coordinates
have a zero value, the search stops, independently of how many
of the optional constraints have been optimized.
When a suboptimal solution is found, the currently satisfied
optional constraints are turned into required constraints,
ensuring that any better solution found in other parts
of the search tree also satisfies these optional constraints.
This could be refined to imposing that the number of
satisfied optional constraints does not decrease, but
it is not clear at this point whether this would be a better strategy.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>