isl_tab_basic_set_constrained_lexmin: allow disjunctive constraints
A disjunctive constraint is a sequence of optional constraints
(except for possibly the last) that is only considered violated
if all the constraints in the sequence are violated.
If the disjunctive constraint is violated, then all the disjuncts
are enforced in turn until one of them can be made to hold.
If none of the disjuncts can be made to hold, then the final
disjunct is marked failed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>