Merge heuristic callback (#6)
* User-definable cluster merge heuristic

isl_schedule_constraints now accepts a callback that, if provided, will be
called instead of a default merge heuristic in the scheduler.  Clustering is
always refused if no cluster-level schedule could be produced since no fusion
is possible in this case.

Additionally, the callback will be called for every pair of node that remained
in different clusters after the main edge-based pass.  This allows for
max-fuse-like strategies that do not require reuse between loops beeing fused.

* compute wcc schedule instead of component if custom merge callback provided

With the default merge heuristic, the scheduler will not even consider merging
different connected components.  Therefore, no clustering is applied between
components.  Custom heuristic may dictate otherwise, so we need to call it
between connceted components and thus call _wcc_schedule instead of
_component_schedule.

* export isl_schedule_node_n_children