introduce cpp bindings generation (with exceptions)
The design of these bindings is essentially the same as
that of the C++ bindings without exceptions, except that
mechanisms that were introduced to avoid exceptions are
not needed.
In particular, no isl::boolean or isl::stat needs to be introduced.
Instead, exceptional results (negative isl_bool and isl_stat,
but also NULL pointer results) cause an exception to be thrown.

Exceptions thrown by C++ callbacks are caught and rethrown,
just like in the Python bindings.

Additionally, all arguments of methods are checked for validity.
This ensures that a valid isl_ctx is available when
an exception needs to be thrown based on an error that occurred
during the execution of the C function.
The exception class derives from std::exception rather than std::logic_error
or std::runtime_error because only some types of isl exceptions
should be considered as logic errors, while some other can
be considered as runtime errors.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>