lightweight isl::id (#18)
All the lifetime management is done by the C side.  Therefore, the deleter must
be implicitly convertible to a C function pointer (no capturing lambdas or
binders).  By default, the deleter does nothing to be consistent with the C
interface.  We may additionally provide a template constructor variant that
provides a lambda with "delete static_cast<T*>(v)" inside.