generate isl::managed_copy functions (#20)
This function avoids the need to write

isl::manage(isl_*_copy(object))

by replacing it with

isl::managed_copy(object)

when the user wants to keep the C object alive after the C++ wrapper is
deleted.  It also allows the user to ignore the type of the C object
thanks to overloading.