Fix "space points to invalid node" and memory management (#22)
* avoid dangling else

* scheduler: graph_find_compressed_node: ensure the node belongs to the graph

graph_find_compressed_node checks if there is a node in the isl_sched_graph
that has the given space. If not, it assumes that the space corresponds to a
compressed node, and that the usr pointer of the space tuple_id points to the
original node. However, the compressed tuple_id is only constructed when the
graph is created from the schedule constraints. When a subgraph is extracted by
extract_sub_graph, new nodes are allocated for the subgraph, and their spaces,
hulls, compression and decompression affs are copied. Therefore, in a subgraph,
the usr pointer of a tuple_id of a node space points to the isl_sched_node of
the parent graph rather than to the isl_sched_node of the subgraph.
graph_find_compressed_node may fail if called on an extracted subgraph with a
space of a compressed node, because it checks whether the node it found belongs
to the subgraph.

Instead of taking the node stored in tuple_id of the given space, take its
space and lookup the node that lives in that space in the subgraph.

* scheduler: only cosntruct merge_callback arguments when necessary

If the callback is not set, the arguments will not be consumed by it and will
leak.