isl_ilp_region: add has_non_zero field
In upcoming commits, the isl_ilp_region structure will be changed
to support other kinds of constraints, in which case a constraint
region may not impose any non-zero constraint.

The mechanism that was already being used
for indicating that there is no non-zero constraint
was to create a non-zero constraint with zero rows.
However, the more natural interpretation of such a constraint
is as one that cannot be satisfied since at least one of
the linear combinations should be non-zero, while there are
no linear combinations.

Add an explicit field to indicate whether there is any non-zero constraint
and interpret a non-zero constraint with zero rows in the more natural
(if somewhat useless) way.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>