isl_edge_type: split isl_edge_last into isl_edge_last_sc and isl_edge_last_table
The value isl_edge_last is used both as the last of the constraints
represented by an isl_union_map in isl_schedule_constraints and
as the last of the edge types for which there is an edge table
in isl_sched_graph.
Currently, these two groups are exactly the same.
However, the representation for inter-statement consecutivity constraints
will not be represented by an isl_union_map, while it will have
an edge table in isl_sched_graph.
The value isl_edge_last therefore needs to be decoupled into
(potentially) separate values for the two use cases.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>