isl_multi_*: store pieces in a union
In particular, only consider the field representing the pieces
to be valid when the number of base expressions is greater than zero.
Since the field is only used to access a base expression,
it is never used when the number of base expression is zero.
Making this condition more explicit allows the field
to overlap with the domain of a zero-dimensional expression,
which will be introduced in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>