cpp: add conversion to bool
In particular, add a conversion operator that converts
any C++ type derived from an isl object type to bool,
returning true if the object represents a valid reference
to an isl object.

The introduction of such an operator has been considered before,
but it was postponed because there are some situations where
this operator might cause confusion.

The C++ bindings without exceptions have an isl::boolean type
that behaves differently when converted to bool.  In particular,
invalid isl::boolean objects print an error message when they
are converted to bool.  The counterargument
is that isl::boolean does not correspond to an isl object type
and therefore should be allowed to be behave differently.
Moreover, the C++ bindings with exceptions do not have such a type.

It may be confusing that an isl::val object equal to 0
converts to true rather than false (which is what the integer 0
would convert to).  However, the user should not take isl::val
to be a drop-in replacement for an integer type, since it can
also represent rational, infinite, and non-values.
This also means that no implicit conversion to integers
will be added.

An is_null method was added instead, which does not
cause any such confusion.

However, in an upcoming commit, the isl_schedule_node subtypes
will be exposed as subclasses of isl::schedule_node and then
it will be convenient to introduce an as() method that checks
if a node is of a given subtype and, if so, casts it to
the corresponding subclass, with a NULL return value indicating
that the node is not of the given subtype.
In this context, it is convenient to have an automatic conversion
to bool in order to be able to write

    if (auto filter = node.as<isl::schedule_node_filter>()) {
        /* operations of filter node */
    }

instead of

    if (node.isa<isl::schedule_node_filter>()) {
        auto filter = isl::schedule_node_filter(node);
        /* operations of filter node */
    }

Since the conversion operator is only meant for contextual conversion,
it is made explicit.

Requested-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>