interface: add support for callbacks not taking arguments
The current generator tacitly assumes that all callbacks
take their arguments.  Make this assumption more explicit and
add support for the case where a callback does not take
its arguments.
This will be used when support for callbacks returning isl_bool
is added, which typically do not take their arguments.

Unfortunately, the memory management annotations of
arguments to functions pointers are not recorded by clang,
so they are not available for figuring out whether
a callback takes its arguments.
Simply continue to assume that it does for now.

As a side effect of this change, parentheses are now placed
around take arguments in the Python callback.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>