scheduler: optionally prevent skewing and scaling
Some applications prefer schedules that do not perform
any skewing (or scaling).  Add an option that bounds
the sum of the absolute values of the coefficients to one
for each statement.
Like all other bounds on coefficients, this option
is ignored by the Feautrier scheduler.

It would be possible to introduce an option that bounds
this sum by an arbitrary number instead, but this would
not play very well with the incremental scheduler
as preventing the sum from accumulating beyond the bound
over successive cluster mergings would require introducing
extra variables in the ILP problem that encode
the absolute values of the coefficients of the original variables
(which may be linear combinations of the cluster variables).
In the special case of bound one, there is no such accumulation.
(Note that the size of individual coefficients
may also grow beyond the schedule_max_coefficient bound
in the current code base, but this can be solved
without introducing such variables.)

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>