isl_pw_*_eval: do not assume default value is zero
For the current piecewise types with a isl_pw_*_eval,
the default value is zero, such that isl_pw_*_eval should
return 0 on a point outside of the domain.
However, other types such as isl_pw_aff do not have
a default value zero.  Return NaN instead in those cases.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>