drop test_islpp
This test file was getting compiled, but was neither run on "make test"
(actually a top-level "make test" simply fails) nor on CI.
Furthermore, it caused a copy of googletest to be downloaded
at configure time.
Remove it.

Eventually, the differences with mainline should either be upstreamed
or moved to the user of this module.
Any generic test in the removed test_islpp should then move to mainline isl.
Any specific test should move to the user.