use generated isl_id bindings
The original motivation for manually defining an isl::id binding
appears to be the handling of the user pointer.
Since access to the user pointer has been removed,
it seems more appropriate to use automatically generated bindings.

Some functionality is removed by this change.
In particular, isl_id_get_hash is not exported anymore.
The original export changed the result type and
is therefore not trivial to carry over.
The equality operator is also no longer exported,
just like there is no equality operator for
any other exported type, even though it can be implemented differently
from those for other types.
For now, the user is expected to use the C interface
for this functionality.