export all lists in the same way
That is, export them in the same way as isl_union_set_list.
The main purpose of the other way of exporting using isl::list
appears to be the ability to support range-based for loops,
but that can be implemented in 20 lines on top of isl::*_list and
therefore does not warrant all the extra (partially undocumented) code.

This essentially reverts isl-0.18-835-g893e7b68e3 (Lists support (#27),
Thu Sep 7 21:49:04 2017 +0200).