isl_ast_node_print: mark coincident for loops
In particular, add a "coincident" comment at the end
of the line.
The comment is added in print_body_c because this is
where the decision is made of whether to print a block and
the comment needs to be added after the opening brace of the block.

The comment is placed on the same line for compactness and
to make it more obvious that it is a property of a particular for-node
rather than a comment generated from a mark-node.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>