add isl_pw_*_involves_param_id
These functions allow a user to check whether a function involves
a parameter without exposing the fact that, internally,
the parameter can also be identified by its position.

Internally, these functions use an isl_pw_*_find_dim_by_id,
but this is not exposed because the isl_*_find_dim_by_* functions
in general have a bad interface, returning -1 both in case
of an error and in case of not finding the dimension.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>