CMakeLists.txt: drop GMP_INCLUDE_DIRS from executable target_include_directories
The path containing the GMP headers may also contain system isl headers,
so it should not be added before the local include paths.
Those are not explicitly set for the executables, apparently
relying on them getting inherited from the isl library
target_include_directories.  Do the same for GMP_INCLUDE_DIRS,
as it is added there after the local include paths.