split up schedule_max_coefficient into schedule_max_{var,param}_coefficient
The schedule_max_coefficient was introduced in isl-0.09-70-g2b54bb607
(scheduler: allow to bound the coefficients in the calculated schedule,
Sun Feb 19 14:05:50 2012 +0100) to bound the schedule coefficients of
both the parameters and the variables based on the observation
that bounding these coefficients can have a significant impact
on the performance of the scheduler.  However, it is not clear
from the commit message whether only bounding the coefficients
of the variables would have been sufficient.
Moreover, the scheduler has been improved significantly since then and
the original observation may no longer apply.
In any case, the effect of variable and parameter coefficients
is very different and it does not seem right to bound them
using the same option.  If anything, the parameter coefficients
are expected to behave more like the constant terms.

The schedule_max_coefficient option is currently not strictly enforced
by the incremental scheduler.  This will be fixed in an upcoming commit,
but it only seems useful to do this for the variable coefficients,
especially because the same infrastructure will be reused later
to optionally only consider non-negative variable coefficients.
There is no need to do the latter for parameter coefficients because
they are already forced to be non-negative since isl-0.16.1-214-g99f23962e
(schedule construction: use only positive parameter coefficients,
Fri Apr 1 16:51:07 2016 +0200).

Split up the schedule_max_coefficient option into
two options, schedule_max_var_coefficient and schedule_max_param_coefficient.
This will allow the schedule_max_var_coefficient option
to be enforced by the incremental scheduler without necessarily
also having to do the same with the schedule_max_param_coefficient option.

For backward compatibility, the old option is kept and whenever
it is set, both schedule_max_var_coefficient and
schedule_max_param_coefficient are set to the same specified value.
The value set by isl_options_set_schedule_max_coefficient
is also stored internally such that it can be returned
by isl_options_get_schedule_max_coefficient.

Reviewed-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>