properly take into account max_var_coefficient in incremental scheduler
The incremental scheduler, which first schedules SCCs separately and
then combines these by incrementally scheduling clusters together,
did not specifically take into account the max_var_coefficient option,
which meant that it was getting applied to both the intra-SCC scheduling and
the inter-cluster scheduling.  In the latter case, the constraint
ends up getting applied to the coefficients of the cluster schedule
rather than to the coefficients of the resulting schedule
on the original nodes.

For example, in the new test case, two incremental cluster schedulings
get performed, each introducing a factor of 2.  When combined,
these result in a coefficient 4 on one of the nodes, even
if a bound of 2 is specified by the user.

Properly take into account the max_var_coefficient option
by translating the corresponding constraints on the variable
coefficients in the original graph to the variable coefficients
in the merge graphs.

Note that the max_var_coefficient is usually set to ensure
that the coefficients do not grow too big, where the specified
value is some more or less arbitrary cut-off point.
Still, for large chains of incremental clusterings,
the original behavior could still end up in very large coefficients.
Also, the same infrastructure will later be reused
to ensure non-negative coefficients, when requested by the user.

In principle, the max_par_coefficient option could
undergo a similar treatment, but it seems less important
to keep those coefficients within their specified bounds.

The constraints on the variable coefficients are collected
after the merge graph is constructed.  An alternative approach
would be to add the constraints on the variable coefficients
to the schedule constraints object and to then extract
the constraints from there during the construction
of the merge graph.  This would allow users to also
specify different kinds of constraints on those coefficients.
However, it seems premature now to offer such functionality.
This can always be reconsidered if the need arises.

This commit introduces a call to isl_basic_set_preimage,
which has an awkward interface.  Arguably, the call should
be replaced by some isl_system_preimage, but such a function
would need to get extracted first.

Reviewed-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>