add schedule_nonneg_var_coefficient
This option allows the user to request the generation of a schedule
with only non-negative values for the coefficients of the variables.
(The coefficients of the parameters are already non-negative.)
As with most options that restrict the values of the coefficients,
this option does not apply to the Feautrier scheduler
(which may get called as a backup).

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>