isl_scheduler.c: zero_out_node_coef: use intra_dim_map
That is, instead of constructing a constraint in the LP space directly,
construct an expression in terms of the node coefficients and
then map that to a constraint in the LP space.
This reduces the number of locations that are aware of
the encoding of the variable coefficients in the LP,
making it easier to change the encoding in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>