use autoconf/automake build system on CI
This is the build system used upstream so it has a higher chance
of being maintained.  As an added bonus, more tests get run
automatically.

Do not build the bindings just yet because some functions
got exported in this repo for which there is not yet any
Python bindings support.