schedule_test.sh: test both whole-component and incremental scheduler
Allow the options recorded in the test case to override the choice
between the whole-component and the incremental scheduler in case
the output depends on this choice.  In this case, the same test
will simply get performed twice.

Since the same test case is now run twice, print
the effective command line arguments in case of failure.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>