cpp_generator::print_method_impl: do not assume return type in case of callback
Ever since it was introduced in isl-0.18-599-g7d3c260fb (cpp: support
methods and constructors, Thu Apr 13 21:42:03 2017 +0200),
cpp_generator::print_method_impl has special handling of the return type
of a function that has a callback, apparently assuming this return type
is always isl_stat.  Explicitly check the return type instead.

When support for callbacks returning isl_bool was added,
the same code also got reused for functions with a callback
that return isl_bool, so check for this type as well.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>