add isl_multi_*_set_range_tuple_id convenience functions
The interface of isl_multi_*_set_tuple_id is not well designed
because it only works for some values of the "type" argument.
Moreover, the value in theory depends on whether
the function lives in a set (isl_dim_set) or a map space (isl_dim_out).
In practice, isl_dim_set is equal to isl_dim_out, but this knowledge
should only be used inside isl.
Provide isl_multi_*_set_range_tuple_id alternatives
to avoid these issues.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>