add isl_space_set_set_tuple_id convenience function
The interface of isl_space_set_tuple_id is not well designed
because it only works for some values of the "type" argument.
Moreover, the value of this argument depends on whether
the first argument is a set or a map space.
Provide a isl_space_set_set_tuple_id alternative to avoid these issues.
In particular, this function should only be called on set spaces.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>