add isl_multi_aff_wrapped_range_map
This function is similar to isl_multi_aff_range_map, but while
isl_multi_aff_range_map takes a map space A -> B and constructs
the function [A -> B] -> B, isl_multi_aff_wrapped_range_map
takes a set space N[A -> B] and construct the function N[A -> B] -> B.
That is, isl_multi_aff_wrapped_range_map allows the user
to prescribe the identifier of the domain of the result.

isl_set_wrapped_domain_map performs a similar operation on sets.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>