document effect of max_coefficient in incremental scheduler
The incremental scheduler, which first schedules SCCs separately and
then combines these by incrementally scheduling clusters together,
does not specifically take into account the max_coefficient option,
which means that it is getting applied to both the intra-SCC scheduling and
the inter-cluster scheduling.  In the latter case, the constraint
ends up getting applied to the coefficients of the cluster schedule
rather than to the coefficients of the resulting schedule
on the original nodes.

This will get fixed (to some extent) in an upcoming commit.
For now, simply document the behavior.

Requested-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>