fix merge callback fallout
The introduction of merge callbacks in isl-0.18-816-g3d84f765 (Merge
heuristic callback (#6), Tue Jul 25 13:21:55 2017 +0200) broke
the handling of maximize coincidence in the absence of merge callbacks.
The commit is probably going to be reverted in the near future,
but in the mean time, the fix is needed to be able to merge master,
which introduces a new test case that breaks.